Change structure of returned data
I like this a little better, since we can add more
resources other than memory later.

Also pass the notebookapp to plugins, rather than
specific config. They can thus use their own
configurables
